small fixes

+++ lex/sqs_purge/fulfillment.py
@@ -1,10 +1,12 @@
 from lex.helpers import get_target, get_aws_client, aws_manager_decorator
-from lex.responses import fulfill
+from lex.responses import fulfill, get_slot
 
 
 @aws_manager_decorator
 def handler(event, aws_config):
     target = get_target(event)
+    if not target:
+        return get_slot(event, 'target', 'missing_target')
     sqs_client = get_aws_client('sqs', aws_config)
 
     sqs_client.purge_queue(QueueUrl=target)
